Refactoring Shipment query.
