+++ src/Action/Shipment/BuyShipmentLabelCommand.php
@@ -1,6 +1,7 @@
 <?php
 namespace inklabs\kommerce\Action\Shipment;
 
+use inklabs\kommerce\EntityDTO\OrderItemQtyDTO;
 use inklabs\kommerce\Lib\Command\CommandInterface;
 
 final class BuyShipmentLabelCommand implements CommandInterface
