+++ src/Action/Shipment/Query/GetShipmentRatesResponseInterface.php
@@ -1,0 +1,9 @@
+<?php
+namespace inklabs\kommerce\Action\Shipment\Query;
+
+use inklabs\kommerce\EntityDTO\ShipmentRateDTO;
+
+interface GetShipmentRatesResponseInterface
+{
+    public function addShipmentRateDTO(ShipmentRateDTO $shipmentRateDTO);
+}
