+++ src/Service/OrderService.php
@@ -1,7 +1,7 @@
 <?php
 namespace inklabs\kommerce\Service;
 
-use inklabs\kommerce\Action\Shipment\OrderItemQtyDTO;
+use inklabs\kommerce\EntityDTO\OrderItemQtyDTO;
 use inklabs\kommerce\Entity\Cart;
 use inklabs\kommerce\Entity\CreditCard;
 use inklabs\kommerce\Entity\CreditPayment;
@@ -131,7 +131,7 @@ class OrderService implements OrderServiceInterface
 
     /**
      * @param int $orderId
-     * @param OrderItemQtyDTO $orderItemQtyDTO
+     * @param \inklabs\kommerce\EntityDTO\OrderItemQtyDTO $orderItemQtyDTO
      * @param string $comment
      * @param int $shipmentCarrierTypeId
      * @param string $trackingCode
@@ -153,7 +153,7 @@ class OrderService implements OrderServiceInterface
 
     /**
      * @param string $comment
-     * @param OrderItemQtyDTO $orderItemQtyDTO
+     * @param \inklabs\kommerce\EntityDTO\OrderItemQtyDTO $orderItemQtyDTO
      * @param ShipmentTracker $shipmentTracker
      * @param Order $order
      */
